fixed mistakes from review
