+++ src/components/OverviewPage.jsx
@@ -2,12 +2,16 @@ import React, { useEffect } from 'react'
 
 import { Container, Text } from '../styled/OverviewPage'
 
-const OverviewPage = ({ overview }) => {
-  useEffect(() => {}, [overview])
+const OverviewPage = ({overview}) => {
+  
+  useEffect(() => {
+    console.log(overview);
+  }, [overview])
+
   return (
     <Container>
       <Text>Overview</Text>
-      {Object.values(overview).map((description, index) => (
+      {overview.map((description, index) => (
         <div key={index}>
           <h3>{Object.values(description)}</h3>
         </div>
