+++ src/components/SearchBar.jsx
@@ -2,18 +2,16 @@ import React from 'react'
 
 import { Search } from '../styled/SearchBar'
 
-const SearchBar = ({handleChange}) => {
-  return (
-    <Search>
-      <input
-        className="form-control me-2"
-        type="search"
-        placeholder="Search"
-        aria-label="Search"
-        onChange={handleChange}
-      />
-    </Search>
-  )
-}
+const SearchBar = ({ handleChange }) => (
+  <Search>
+    <input
+      className="form-control me-2"
+      type="search"
+      placeholder="Search"
+      aria-label="Search"
+      onChange={handleChange}
+    />
+  </Search>
+)
 
 export default SearchBar
