+++ src/components/UsersList.jsx
@@ -6,28 +6,29 @@ import { Container, Text } from '../styled/UsersList'
 import SearchBar from './SearchBar'
 
 const UsersList = ({ employees, getOverview }) => {
+
   const [searchTerm, setSearchTerm] = React.useState('')
   const [searchResults, setSearchResults] = React.useState([])
 
   const handleChange = (event) => {
     setSearchTerm(event.target.value)
-  }
+  };
 
   useEffect(() => {
     const results = employees.filter((employee) =>
       employee.toLowerCase().includes(searchTerm)
     )
     setSearchResults(results)
-  }, [searchTerm, employees, getOverview])
+  }, [searchTerm, employees])
 
   return (
     <>
-      <SearchBar handleChange={handleChange} />
+     <SearchBar handleChange={handleChange} />
       <Container className='row'>
         {searchResults.length > 0 ? (
           searchResults.map((employee, index) => (
-            <Card style={{ width: '18rem', margin: `0.5rem`}}>
-              <Card.Body key={index} >
+            <Card key={index} style={{ width: '18rem', margin: `0.5rem`}}>
+              <Card.Body >
                 <Card.Title>{employee}</Card.Title>
                 <Link
                   variant="primary"
