+++ src/redux/store.js
@@ -1,8 +1,13 @@
-import { applyMiddleware, compose, createStore } from 'redux';
+import { applyMiddleware, compose, createStore , combineReducers } from 'redux';
 import createSagaMiddleware from 'redux-saga'
 
-import reducer from './reducers'
-import {watchedEmployees, watchedOverviews} from './sagas'
+import { watchedOverviews, overview } from './ducks/overviews'
+import { watchedEmployees, employee } from './ducks/employees'
+
+const reducer = combineReducers({
+  employees: employee,
+  overviews: overview
+})
 
 const sagaMiddleware = createSagaMiddleware()
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
