--- src/redux/types.js
@@ -1,11 +1,0 @@
-const SET_EMPLOYEE = 'SET_EMPLOYEE'
-const SET_OVERVIEW = 'SET_OVERVIEW'
-const LOAD_EMPLOYEE = 'LOAD_EMPLOYEE'
-const LOAD_OVERVIEW = 'LOAD_OVERVIEW'
-
-export {
-  SET_EMPLOYEE,
-  SET_OVERVIEW,
-  LOAD_EMPLOYEE,
-  LOAD_OVERVIEW,
-}
