+++ src/styled/SearchBar.js
